# swagger.yaml
openapi: 3.0.2
info:
  title: BarPank
  version: 1.0.0
  description: ''
paths:
  /users:
    summary: Path used to manage the list of users.
    description: >-
      The REST endpoint/path used to list and create zero or more `user` entities.  This path contains a
      `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `user` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
        required: true
      tags:
        - users
      responses:
        '201':
          description: Successful response.
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                parameter missing:
                  value:
                    error: >-
                      user validation failed: (?<parameterName>[^:]+): Path
                      (?<parameterNameAgain>[^]+) is required.(, ([^:]+): Path ([^]+) is
                      required.){0,2}
          description: Parameter missing
        '409':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                username already exists:
                  value:
                    error: Username already exists
          description: username already exists
        '422':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Parameter(s) value too short/long:
                  value: "{\r\n    \"error\": \"Path (?<parameterName>.+) ((?<sampleValue>.+)) is (longer|shorte) than the (maximum|minimum) allowed length ((?<requiredLength>\\d+))\"\r\n}"
          description: "Parameter(s) value too short/long\r\n"
      operationId: createuser
      summary: Create a user
      description: Creates a new instance of a `user`.
  /users/current:
    get:
      tags:
        - users
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/user'
              examples:
                GET /users/current response:
                  value:
                    accounts:
                      -
                        account_number: fgb2tqxwc0zx
                        balance: 10000
                        currency: EUR
                        name: Main
                        userId: 616564d768c0f084cd704a38
                    name: Tere11
                    username: Test
          description: Successful request
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid authorization format:
                  value:
                    error: Invalid authorization format
          description: bad request
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Missing authorization:
                  value:
                    error: Missing authorization
                Invalid token:
                  value:
                    error: Invalid token
          description: unauthorized
      security:
        -
          Bearer: []
      summary: Get active user data
      description: Returns active `user`data
  /sessions:
    summary: Path used to manage the list of sessions.
    description: >-
      The REST endpoint/path used to list and create zero or more `session` entities.  This path
      contains a `GET` and `POST` operation to perform the list and create tasks, respectively.
    post:
      requestBody:
        description: A new `session` to be created.
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/user'
            examples:
              POST /sessions request:
                value:
                  username: anna
                  password: qwerty1234
        required: true
      tags:
        - sessions
      responses:
        '200':
          description: successful login
        '400':
          content:
            application/json:
              examples:
                Required parameter missing:
                  value:
                    error: Required parameter missing
          description: Required parameter missing
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Missing authorization:
                  value:
                    error: Missing authorization
          description: unauthorized
      operationId: createsession
      summary: Create a session
      description: Creates a new instance of a `session`.
    delete:
      tags:
        - sessions
      responses:
        '204':
          description: Session successfully deleted
        '400':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                Invalid authorisation format:
                  value:
                    error: Invalid authorisation format
          description: Invalid authorisation format
        '401':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/error'
              examples:
                invalid token:
                  value:
                    error: invalid token
                Missing authorization:
                  value:
                    error: Missing authorization
          description: invalid token
      security:
        -
          Bearer: []
      description: "Deletes given `session` with an id specified in authorization header\r\n"
components:
  schemas:
    user:
      title: Root Type for user
      description: ''
      type: object
      properties:
        username:
          type: string
        password:
          type: string
        name:
          type: string
      example:
        username: miki
        password: plutoonium123
        name: Miki Hiir
    error:
      title: Root Type for error
      description: ''
      type: object
      properties:
        error:
          type: string
      example:
        error: some text
    session:
      title: Root Type for session
      description: ''
      type: object
      properties:
        token:
          type: string
      example:
        token: 6162ea8470ba6a0cc3cabdbd
  securitySchemes:
    Bearer:
      scheme: bearer
      type: http
tags:
  -
    name: users
    description: ''
  -
    name: sessions
    description: ''